+++ main.ts
@@ -1,4 +1,4 @@
-const { app, BrowserWindow, Menu } = require("electron");
+const { app, BrowserWindow, Menu, dialog } = require("electron");
 const url = require("url");
 const path = require("path");
 
@@ -27,6 +27,19 @@ function createWindow() {
       slashes: true,
     })
   );
+
+  mainWindow.on("close", function (e) {
+    var choice = dialog.showMessageBoxSync(this, {
+      type: "question",
+      buttons: ["Yes", "No"],
+      title: "Confirm",
+      message: "Are you sure you want to quit?",
+      detail: "All you unsaved progress will be lost!",
+    });
+    if (choice == 1) {
+      e.preventDefault();
+    }
+  });
 
   mainWindow.on("closed", function () {
     mainWindow = null;