+++ src/app/core/services/event-emitter/event-emitter.service.ts
@@ -11,7 +11,6 @@ export class EventEmitterService {
   constructor() {}
 
   onSaveFileTrigger() {
-    console.log("EVENT EMITTED");
     this.invokeSaveFile.emit();
   }
 }